whole heap of changes: new css formatting. new search method. refactor live_search. added testing in the readme. fixed LEFT JOIN. Made franchisess list in draft order. Added default in players schema.
git-svn-id: svn+ssh://www.mtbcalendar.com/home/kpepple/svn/draftday/trunk@17 45e1ab1c-cbb3-db11-9a89-001422107726

+++ db/schema.rb
@@ -26,7 +26,7 @@ ActiveRecord::Schema.define(:version => 6) do
     t.column "K", :integer
     t.column "ERA", :float
     t.column "WHIP", :float
-    t.column "R", :float
+    t.column "R", :integer
     t.column "HR", :integer
     t.column "RBI", :integer
     t.column "SB", :integer
